requirements.txt
# Core GUI Framework
PyQt5>=5.15.0

# Desktop Automation
pyautogui>=0.9.53
pywin32>=305
Pillow>=9.0.0  # Required by pyautogui for screenshots

# Excel Automation
openpyxl>=3.0.10
xlrd>=2.0.1  # For reading older .xls files

# Office Automation (Word, Outlook)
python-docx>=0.8.11

# SharePoint and Microsoft Graph API
Office365-REST-Python-Client>=2.3.0
msal>=1.20.0
requests>=2.28.0

# Task Scheduling
APScheduler>=3.10.0

# Configuration Management
PyYAML>=6.0

# Utilities
python-dateutil>=2.8.2

# Packaging
PyInstaller>=5.0
